🏗️ Refactor: 상세 정보 fetch 로직 커스텀 훅으로 관리
